+++ services/common/utils/middlewares.js
@@ -22,11 +22,22 @@ const needAccessToken = (accessTokenSecret) => {
 
     try {
       const decoded = jwt.verify(accessToken, accessTokenSecret);
-      if (!decoded.account) {
+      if (decoded.account) {
+        // user type UserServicePrincipal
+        req.user = {
+          type: 'user',
+          email: decoded.account.email,
+        };
+      } else if (decoded.application) {
+        // user type ApplicationServicePrincipal
+        req.user = {
+          type: 'application',
+          clientId: decoded.application.clientId,
+        };
+      } else {
         logger.warn('accessToken is invalid');
         return res.sendStatus(401);
       }
-      req.user = decoded.account;
     } catch (error) {
       logger.warn(error);
       return res.sendStatus(401);
@@ -43,10 +54,29 @@ const checkOrganization = () => {
       return next();
     }
 
-    // for the currennt user, add all subscribed organizations in request object
-    req.realms = (
-      await Realm.find({ members: { $elemMatch: { email: req.user.email } } })
-    ).map((realm) => {
+    let realms;
+    switch (req.user.type) {
+      case 'user':
+        // for the current user, add all subscribed organizations in request object
+        realms = await Realm.find({
+          members: { $elemMatch: { email: req.user.email } },
+        });
+        break;
+      case 'application':
+        // for the current application access, add only the associated realm
+        const realm = await Realm.findOne({
+          applications: { $elemMatch: { clientId: req.user.clientId } },
+        });
+        realms = realm ? [realm] : [];
+        break;
+      default:
+        logger.error(
+          'checkOrganization: Invalid request received: neither user nor application'
+        );
+        return res.sendStatus(500);
+    }
+    // transform realms to objects
+    req.realms = realms.map((realm) => {
       const r = realm.toObject();
       r._id = r._id.toString();
       return r;
@@ -79,6 +109,24 @@ const checkOrganization = () => {
     // current user is not a member of the organization
     if (!req.realms.find(({ _id }) => _id === req.realm._id)) {
       logger.warn('current user is not a member of the organization');
+      return res.sendStatus(404);
+    }
+
+    // resolve the role for the current realm
+    switch (req.user.type) {
+      case 'user':
+        req.user.role = req.realm.members.find(
+          ({ email }) => email === req.user.email
+        )?.role;
+        break;
+      case 'application':
+        req.user.role = req.realm.applications.find(
+          ({ clientId }) => clientId === req.user.clientId
+        )?.role;
+        break;
+    }
+    if (!req.user.role) {
+      logger.warn('current user could no be found within realm');
       return res.sendStatus(404);
     }